Move TimeseriesList, PeriodseriesList to energyquantified.data
