+++ energyquantified/api/instances.py
@@ -1,4 +1,3 @@
-from ..utils import Page, dict_to_str
 from .base import BaseAPI
 
 from ..exceptions import ValidationError
