+++ energyquantified/data/__init__.py
@@ -1,17 +1,31 @@
-from .timeseries import Timeseries, ValueType, Value, ScenariosValue, MeanScenariosValue
-from .periodseries import Periodseries, Period, CapacityPeriod
+from .timeseries import (
+    Timeseries,
+    TimeseriesList,
+    ValueType,
+    Value,
+    ScenariosValue,
+    MeanScenariosValue
+)
+from .periodseries import (
+    Periodseries,
+    PeriodseriesList,
+    Period,
+    CapacityPeriod
+)
 from .ohlc import Product, OHLC, OHLCList
 
 
 __all__ = [
     # Time series
     "Timeseries",
+    "TimeseriesList",
     "ValueType",
     "Value",
     "ScenariosValue",
     "MeanScenariosValue",
     # Period series
     "Periodseries",
+    "PeriodseriesList",
     "Period",
     "CapacityPeriod",
     # OHLC
