+++ energyquantified/parser/periodseries.py
@@ -1,9 +1,8 @@
 from dateutil import parser
 
-from ..data import Period, CapacityPeriod, Periodseries
+from ..data import Period, CapacityPeriod, Periodseries, PeriodseriesList
 from ..exceptions import ParseException
 from ..time import to_timezone
-from ..utils import PeriodseriesList
 from .metadata import parse_curve, parse_instance, parse_resolution
